- fix potential little bug
git-svn-id: svn+ssh://siuc04/sc02a1/users/ccpsoeta/svnRep/ukb/Main@44 41fde781-4c32-0410-9b47-839176a67e93

+++ disambGraph.cc
@@ -356,7 +356,7 @@ void hits_init_ranks(DisambG & g,
 		     HProp hProp, AProp aProp, size_t n) {
 
   assert(n);
-  float init_v = sqrt(n) / n;
+  float init_v = sqrt(n) / static_cast<float>(n);
 
   //graph_traits<DisambG>::vertex_iterator vit, end;
   //tie(vit, end) =  vertices(g);
